requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
jinja2
psycopg2-binary
aiofiles
python-dotenv
websockets
